redux-todos: Move actions & constants to own file
Let's make sure we do not use problematic cyclic dependencies.
