+++ packages/redux-todos/src/actions.js
@@ -1,0 +1,10 @@
+import { ADD_TODO } from './constants';
+
+export const addTodo = todo => ({
+  type: ADD_TODO,
+  todo,
+});
+
+export default {
+  addTodo,
+};
