+++ packages/redux-todos/src/constants.js
@@ -1,0 +1,5 @@
+export const ADD_TODO = 'ADD_TODO';
+
+export default {
+  ADD_TODO,
+}
