+++ packages/redux-todos/src/index.test.js
@@ -1,10 +1,12 @@
-import todos, { actions, types } from './';
+import reducer  from './';
+import constants from './constants';
+import actions from './actions';
 
 describe('actions', () => {
   it('should create an action to add a todo', () => {
     const todo = 'foo';
     const expectedAction = {
-      type: types.ADD_TODO,
+      type: constants.ADD_TODO,
       todo,
     };
     expect(actions.addTodo(todo)).toEqual(expectedAction);
@@ -14,8 +16,8 @@ describe('actions', () => {
 describe('reducer', () => {
   it('should add a todo', () => {
     const todo = 'a todo';
-    const state = todos(undefined, {});
-    const newState = todos(state, actions.addTodo(todo));
+    const state = reducer(undefined, {});
+    const newState = reducer(state, actions.addTodo(todo));
     const expectedState = [todo];
     expect(newState).toEqual(expectedState);
   });
